+++ sokannonser/rest/model/fields.py
@@ -4,6 +4,7 @@ from sokannonser.repository import taxonomy
 ID = 'id'
 EXTERNAL_ID = 'external_id'
 AD_URL = 'webpage_url'
+LOGO_URL = 'logo_url'
 HEADLINE = 'headline'
 APPLICATION_DEADLINE = 'application_deadline'
 NUMBER_OF_VACANCIES = 'number_of_vacancies'