+++ sokannonser/rest/model/platsannons_results.py
@@ -28,6 +28,15 @@ class AdUrl(fields.Raw):
             return "%s/%s" % (settings.BASE_PB_URL, value)
 
 
+class LogoUrl(fields.Raw):
+    def format(self, value):
+        if settings.BASE_URL[-1] == '/':
+            return "%sad/%s/logo" % (settings.BASE_URL, value)
+        else:
+            return "%s/ad/%s/logo" % (settings.BASE_URL, value)
+
+
+
 taxonomy_item = ns_platsannons.model('JobTechTaxonomyItem', {
     'concept_id': fields.String(),
     'label': fields.String(),
@@ -93,6 +102,7 @@ job_ad = ns_platsannons.model('JobAd', {
     f.ID: fields.String(),
     f.EXTERNAL_ID: fields.String(),
     f.AD_URL: AdUrl(attribute='id'),
+    f.LOGO_URL: LogoUrl(attribute='id'),
     f.HEADLINE: fields.String(),
     f.APPLICATION_DEADLINE: fields.DateTime(),
     f.NUMBER_OF_VACANCIES: fields.Integer(),