+++ sokannonser/settings.py
@@ -27,7 +27,9 @@ RESTPLUS_ERROR_404_HELP = False
 
 
 # Base API URL
-BASE_URL = os.getenv('BASE_URL', 'https://base.url')
+BASE_URL = os.getenv('BASE_URL', 'https://jobsearch.api.jobtechdev.se/')
+BASE_PB_URL = os.getenv('BASE_PB_URL',
+                        'https://www.arbetsformedlingen.se/For-arbetssokande/Platsbanken/annonser/')
 
 # Header parameters
 APIKEY = 'api-key'