+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Changelog Application Jobsearch
 ===============================
 
 ## 1.4.0
-* New ad url field in job ad
+* New webpage url field in job ad
 * Language added as filter in taxonomy search
 
 ## 1.3.0