+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog Application Jobsearch
 ===============================
 
+## 1.4.0
+* New webpage url field in job ad
+* New logo url field in job ad
+* Language added as filter in taxonomy search
+
 ## 1.3.0
 * New parameter: relevance-threshold
 * Introduces new field: score